split oauth api functions
